Move tests file to test folder
