+++ packages/@n8n/agents/src/runtime/__tests__/event-bus.test.ts
@@ -1,4 +1,4 @@
-import { AgentEventBus } from '../runtime/event-bus';
+import { AgentEventBus } from '../event-bus';
 
 describe('AgentEventBus', () => {
 	describe('resetAbort', () => {
