+++ packages/@n8n/agents/src/runtime/__tests__/inmemory-messages.test.ts
@@ -1,5 +1,5 @@
-import { InMemoryMemory } from '../runtime/memory-store';
-import type { AgentDbMessage, AgentMessage, Message } from '../types/sdk/message';
+import type { AgentDbMessage, AgentMessage, Message } from '../../types/sdk/message';
+import { InMemoryMemory } from '../memory-store';
 
 function makeMsg(role: 'user' | 'assistant', text: string): AgentDbMessage {
 	return {
