+++ packages/@n8n/agents/src/runtime/__tests__/message-list.test.ts
@@ -1,6 +1,11 @@
-import { AgentMessageList } from '../runtime/message-list';
-import { isLlmMessage } from '../sdk/message';
-import type { AgentDbMessage, AgentMessage, ContentToolCall, Message } from '../types/sdk/message';
+import { isLlmMessage } from '../../sdk/message';
+import type {
+	AgentDbMessage,
+	AgentMessage,
+	ContentToolCall,
+	Message,
+} from '../../types/sdk/message';
+import { AgentMessageList } from '../message-list';
 
 function makeUserMsg(text: string): AgentMessage {
 	return { role: 'user', content: [{ type: 'text', text }] };
