+++ packages/@n8n/agents/src/sdk/__tests__/telemetry.test.ts
@@ -1,6 +1,6 @@
 import type { TelemetryIntegration } from 'ai';
 
-import { Telemetry } from '../sdk/telemetry';
+import { Telemetry } from '../telemetry';
 
 describe('Telemetry builder', () => {
 	it('builds with defaults', async () => {
